CSS improvment (#8)
* sidebar fixed and loading screen

* fixed somethings

* better main page

* add no board handling

* fix delete board problem and get rid of warnings

* get rid of console logs

* sidebar fix for real this time

* new colors

* add looping

* fix modal layering

* fix merge conflict
